fix(package-analyzing): allow deeper levels for source root
Before, the package-analyzer looked for the node_modules folder in only exactly one path step back
to the source files.
I changed this behaviour to use a node-like lookup which looks for the node_modules folder in
the following manner:
./node_modules, ../node_modules, ../../node_modules, etc.

And this allows to put the root of the source (project.paths.root) somewhere deeper in a file-structure.
See Issue #278