fix(bundle): hash filename after loader config added
Resolves: https://github.com/aurelia/cli/issues/465
Previously, a file hash would be generated for the config target before
the loader config was added. This would cause cache busting issues
since a vendor bundle would not have any changes, and therefore have an
identical hash as previous, but other bundles would have a different
hash. The browser cached version of the vendor bundle would now
incorrectly point to a wrong file.
By moving the hashing to after the loader config is added, we now catch
the caching issue, and generate a new (correct) hash.