feat(cli-options): notify user of non-standard environment
Related: https://github.com/aurelia/cli/issues/488
When a user has a custom node environment (`NODE_ENV`), issues can
arise where the CLI tries to get an environment file named from the
`NODE_ENV`. This typically results in a file not found exception with
no real clues as to what went wrong.
This fix will notify the user if the `NODE_ENV` is being used, and if
it is a non-standard environment option (dev, stage, prod)