refactor(bundler): uses uglify-es in place of uglify-js
With better support of ES6+ in modern browsers and the introduction
of babel-preset-env it's possible to serve ES6 code to clients able
to consume it. This change uses uglify-es in place of ugliify-js to
handle minification that supports ES6.

uglify-es is API/CLI compatible with uglify-js@3.
uglify-es is not backwards compatible with uglify-js@2.