fix(cli): default bundler is webpack, so move to position 1
Because index 0 is used as a means of classifying an option as being (Default) when you choose custom in the CLI at present, RequireJS is presented in the UI as the default. However, Webpack is now the default. The fix for this is to move Webpack to the top of the list so the code on line 155 `text += ' (Default)';` applies to Webpack.