fix(bundler): improve compatibility with nodejs global var "global"
A simple global var "global" is created before load all AMD modules. This is a practical way to support nodejs global var "global". In future we will build on top of dumberjs to properly support nodejs global vars "global", "process", and "Buffer" (the last two are still not supported transparently in aurelia-cli bundler). With this fix, @babel/polyfill now works as an AMD module.