refactor(skeleton): remove bluebird, use the promise polyfill in babel/polyfill
This is not a breaking change to existing apps. Existing apps with bluebird are not affected at all. This only affects new app.
The promise polyfill provided by babel is very good, for example, it provides promise.prototype.finally on browsers have native promise but do not provide method finally. It guarantees to normalise browser.