feat(webpack):enable CLI-less webpack build + run
Usage of aurelia-cli is removed from building and running
the webpack apps. This gives developers more control over
webpack, and a way to easily change the configuration,
without aurelia-cli meddling in between.

The commands `au run` or `au build` will not work anymore
in the webpack apps, use `npm start`, or `npm run build`
instead. This commit closes the webpack part of the
the issue #1098.

Conflicts:
	skeleton/webpack/webpack.config.js