Null reference issue run/watch on scss change
I have a scenario where a bundled scss->css files that are not directly referenced by the aurelia app but are loaded dynamically based on user-selected them have references to other scss files.  When I do an au run and then edit one of these files, which are referenced in an aurelia project cssProcess source array glob to trigger rebuild does, but then the code crashes on rebuild due to a null reference, which turns out to be this.includedIn in the bundled-source.js file.  Doing a simple null check fixes everything for my scenario.